+++ SRC/app/src/main/java/com/example/quokkapuffevents/view/AdminNotifFragAdapter.java
@@ -1,5 +1,6 @@
 package com.example.quokkapuffevents.view;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -40,9 +41,12 @@ public class AdminNotifFragAdapter extends ArrayAdapter<Notif> {
 
         Notif currentNotif = notifList.get(position);
 
+        TextView notifName = (TextView) listItem.findViewById(R.id.notifName);
+        notifName.setText(currentNotif.getMessage());
+
         TextView notifTitles = (TextView) listItem.findViewById(R.id.originUser);
         db.GetUser(currentNotif.getOriginUser(), user -> {
-            notifTitles.setText(user.getFirstName());
+            notifTitles.setText(user.getUserName());
         });
 
         Button deleteButton = listItem.findViewById(R.id.deleteButton);
@@ -50,9 +54,17 @@ public class AdminNotifFragAdapter extends ArrayAdapter<Notif> {
             @Override
             public void onClick(View v) {
                 Notif notif = getItem(position);
-                db.DeleteNotification(notif);
-                notifList.remove(notif);
-                notifyDataSetChanged();
+                new AlertDialog.Builder(getContext())
+                        .setTitle("Delete Notification")
+                        .setMessage("Are you sure you want to delete this notification?")
+                        .setPositiveButton("Delete", (dialog, which) -> {
+                            db.DeleteNotification(notif);
+                            notifList.remove(notif);
+                            notifyDataSetChanged();
+                        })
+                        .setNegativeButton("Cancel", (dialog, which) -> dialog.dismiss())
+                        .show();
+
             }
         });
