+++ SRC/app/src/main/java/com/example/quokkapuffevents/view/AdminUserFragAdapter.java
@@ -1,5 +1,6 @@
 package com.example.quokkapuffevents.view;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -7,6 +8,7 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -45,20 +47,43 @@ public class AdminUserFragAdapter extends ArrayAdapter<User> {
         db = Database.getInstance();
 
         User currentUser = userList.get(position);
-        TextView userTexts = (TextView) listItem.findViewById(R.id.firstName);
-        userTexts.setText(currentUser.getFirstName());
 
-        TextView emailTexts = (TextView) listItem.findViewById(R.id.lastName);
-        emailTexts.setText(currentUser.getLastName());
+        TextView userTexts = (TextView) listItem.findViewById(R.id.UserNameTextView);
+        userTexts.setText(currentUser.getUserName());
+
+        TextView firstAndLastNameText = (TextView) listItem.findViewById(R.id.FirstAndLastNameTextView);
+        String firstLast = currentUser.getFirstName() + " " + currentUser.getLastName();
+        firstAndLastNameText.setText(firstLast);
+
+        ImageView accountImg = (ImageView) listItem.findViewById(R.id.imageView);
+
+        Integer type = currentUser.getAccountType();
+        if (type == -1) { // Admin
+            accountImg.setImageResource(R.drawable.admin_icon);
+        }
+        else if (type == 0) { // Entrant
+            accountImg.setImageResource(R.drawable.entrant_icon);
+        }
+        else if (type == 1) { // Organizer
+            accountImg.setImageResource(R.drawable.organizer_icon);
+        }
 
         Button deleteButton = listItem.findViewById(R.id.deleteButton);
         deleteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 User user = getItem(position);
-                db.DeleteUser(user);
-                userList.remove(user);
-                notifyDataSetChanged();
+
+                new AlertDialog.Builder(getContext())
+                        .setTitle("Delete User")
+                        .setMessage("Are you sure you want to delete this user?")
+                        .setPositiveButton("Delete", (dialog, which) -> {
+                            db.DeleteUser(user);
+                            userList.remove(user);
+                            notifyDataSetChanged();
+                        })
+                        .setNegativeButton("Cancel", (dialog, which) -> dialog.dismiss())
+                        .show();
             }
         });
